+++ src/libs/decorators/number/minValue.mjs
@@ -36,10 +36,11 @@ function _featureMinValue(params, value, adjust = false)
  * adjust
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(values.adjusted >= params.value)
 	{
@@ -51,5 +52,5 @@ function _adjust(params, values)
 		return false;
 	}
 
-	AdjusterError.raise(CAUSE.MIN_VALUE, values);
+	AdjusterError.raise(CAUSE.MIN_VALUE, values, keyStack);
 }