+++ src/libs/decorators/numericString/checksum.mjs
@@ -35,10 +35,11 @@ function _featureChecksum(params, algorithm)
  * adjust
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(!params.flag)
 	{
@@ -49,7 +50,7 @@ function _adjust(params, values)
 		return false;
 	}
 
-	AdjusterError.raise(CAUSE.CHECKSUM, values);
+	AdjusterError.raise(CAUSE.CHECKSUM, values, keyStack);
 }
 
 /**