+++ src/libs/decorators/numericString/pattern.mjs
@@ -11,15 +11,16 @@ export default AdjusterBase.decoratorBuilder(_adjust)
  * adjust
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(REGEXP.test(values.adjusted))
 	{
 		return false;
 	}
 
-	AdjusterError.raise(CAUSE.PATTERN, values);
+	AdjusterError.raise(CAUSE.PATTERN, values, keyStack);
 }