+++ src/libs/decorators/object/constraints.mjs
@@ -1,4 +1,4 @@
-import adjust from "../../adjust";
+import adjustObject from "../../adjustObject";
 import AdjusterBase from "../../AdjusterBase";
 
 export default AdjusterBase.decoratorBuilder(_adjust)
@@ -33,16 +33,17 @@ function _featureConstraints(params, constraints)
  * adjuster
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(params.constraints === null)
 	{
 		return false;
 	}
 
-	values.adjusted = adjust(values.adjusted, params.constraints);
+	values.adjusted = adjustObject(values.adjusted, params.constraints, AdjusterBase.onErrorDefault, keyStack);
 	return false;
 }