+++ src/libs/decorators/only.mjs
@@ -35,10 +35,11 @@ function _featureOnly(params, ...values)
  * adjust
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(!params.flag)
 	{
@@ -49,5 +50,5 @@ function _adjust(params, values)
 		return true;
 	}
 
-	AdjusterError.raise(CAUSE.ONLY, values);
+	AdjusterError.raise(CAUSE.ONLY, values, keyStack);
 }