+++ src/libs/decorators/string/maxLength.mjs
@@ -37,10 +37,11 @@ function _featureMaxLength(params, length, adjust = false)
  * adjust
  * @param {Object} params parameters
  * @param {AdjusterBase.VALUES} values original / adjusted values
+ * @param {(string|number)[]} keyStack path to key that caused error
  * @returns {boolean} end adjustment
  * @throws {AdjusterError}
  */
-function _adjust(params, values)
+function _adjust(params, values, keyStack)
 {
 	if(!params.flag)
 	{
@@ -57,5 +58,5 @@ function _adjust(params, values)
 		return false;
 	}
 
-	AdjusterError.raise(CAUSE.MAX_LENGTH, values);
+	AdjusterError.raise(CAUSE.MAX_LENGTH, values, keyStack);
 }