+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `adjuster.string().strict()`
 * `adjuster.array()`
 * `adjuster.object()`
+* `err.keyStack` indicates path to key that caused error; for nested object or array
 
 ### Changed
 * cause an error `adjuster.CAUSE.TYPE` instead of  `adjuster.CAUSE.NOT_OBJECT`
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deleted
 * `adjuster.numberArray()` - use `adjuster.array()` instead
 * `adjuster.stringArray()` - use `adjuster.array()` instead
+* `err.key` - use `err.keyStack` instead
 
 ## [0.14.0] - 2018-08-11
 ### Added